Add motion planning options inside circle_motion.py. Add environment for collision test.
