+++ jsk_uav_forest_simulation/gazebo_model/world/forest_collision.world
@@ -1,0 +1,165 @@
+<?xml version="1.0" ?>
+<sdf version='1.4'>
+  <world name='default'>
+
+    <physics type='ode'>
+      <max_step_size>0.001</max_step_size>
+      <real_time_factor>1</real_time_factor>
+      <real_time_update_rate>1000</real_time_update_rate>
+      <gravity>0 0 -9.8</gravity>
+    </physics>
+
+    <spherical_coordinates>
+      <surface_model>EARTH_WGS84</surface_model>
+      <latitude_deg>0</latitude_deg>
+      <longitude_deg>0</longitude_deg>
+      <elevation>0</elevation>
+      <heading_deg>0</heading_deg>
+    </spherical_coordinates>
+
+
+    <!-- A global light source -->
+   <scene>
+      <sky>
+        <clouds>
+          <speed>12</speed>
+        </clouds>
+      </sky>
+   </scene>
+
+   <include>
+     <uri>model://sun</uri>
+   </include>
+
+   <model name='asphalt_plane'>
+     <static>1</static>
+     <link name='link'>
+       <collision name='collision'>
+         <geometry>
+           <box>
+             <size>20 20 0.1</size>
+           </box>
+         </geometry>
+         <max_contacts>10</max_contacts>
+         <surface>
+           <contact>
+             <ode/>
+           </contact>
+           <bounce/>
+         </surface>
+       </collision>
+       <visual name='visual'>
+         <cast_shadows>0</cast_shadows>
+         <geometry>
+           <box>
+             <size>20 20 0.1</size>
+           </box>
+         </geometry>
+         <material>
+           <script>
+             <uri>model://asphalt_plane/materials/scripts</uri>
+             <uri>model://asphalt_plane/materials/textures</uri>
+             <name>vrc/asphalt</name>
+           </script>
+         </material>
+       </visual>
+       <self_collide>0</self_collide>
+       <kinematic>0</kinematic>
+       <gravity>1</gravity>
+     </link>
+   </model>
+
+    <model name="target_tree" >
+      <pose>0 0 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree1" >
+      <pose>2 2 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree2" >
+      <pose>0 4 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree3" >
+      <pose>-2 -2 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree4" >
+      <pose>-4 0 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree5" >
+      <pose>-4.5 1.5 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <model name="sub_target_tree6" >
+      <pose>-5 4 0 0 0 0</pose>
+      <include>
+        <uri>model://tree</uri>
+      </include>
+    </model>
+
+    <!-- Please check the spefic rule: http://www.lsse.kyutech.ac.jp/~sociorobo/ja/forestdrone2016r -->
+    <model name="red_band">
+      <pose>0 0 0.8 0 0 0</pose>
+      <static>1</static>
+      <link name="band" >
+        <pose>0 0 0 0 0 0</pose>
+        <inertial>
+          <mass>1</mass>
+          <inertia>
+            <ixx>0.145833</ixx>
+            <ixy>0</ixy>
+            <ixz>0</ixz>
+            <iyy>0.145833</iyy>
+            <iyz>0</iyz>
+            <izz>0.125</izz>
+          </inertia>
+          <pose>0 0 0 0 -0 0</pose>
+        </inertial>
+        <collision name="collision">
+          <geometry>
+            <cylinder>
+              <length>0.3</length>
+              <radius>0.11</radius>
+            </cylinder>
+          </geometry>
+        </collision>
+        <visual name="visual">
+          <geometry>
+            <cylinder>
+              <length>0.3</length>
+              <radius>0.11</radius>
+            </cylinder>
+          </geometry>
+          <material>
+            <ambient>1.00 0.0 0.0 1</ambient> <!-- red -->
+            <diffuse>1.00 0.38 0.28 1</diffuse>
+            <specular>0.1 0.1 0.1 1</specular>
+            <emissive>0 0 0 0</emissive>
+          </material>
+        </visual>
+      </link>
+    </model>
+
+  </world>
+</sdf>
