+++ jsk_uav_forest_simulation/launch/forest_simulation.launch
@@ -9,13 +9,16 @@
   <arg name="task_kind" default="2" />
   <arg name="rqt_gui" default="True" />
   <arg name="circle_motion_times" default="1" />
+  <arg name="plannar" default="false" />
 
-  <arg name="start_x" default="-8"/>
-  <arg name="start_y" default="0"/>
-  <arg name="start_z" default="0.5"/>
+  <arg name="start_x" value="-8"/>
+  <arg name="start_y" value="0" unless="$(arg plannar)"/>
+  <arg name="start_y" value="-1" if="$(arg plannar)"/>
+  <arg name="start_z" value="0.5"/>
 
   <include file="$(find gazebo_ros)/launch/empty_world.launch">
-    <arg name="world_name" value="$(find jsk_uav_forest_simulation)/gazebo_model/world/forest.world" />
+    <arg name="world_name" value="$(find jsk_uav_forest_simulation)/gazebo_model/world/forest.world" unless="$(arg plannar)" />
+    <arg name="world_name" value="$(find jsk_uav_forest_simulation)/gazebo_model/world/forest_collision.world" if="$(arg plannar)" />
     <arg name="debug" value="$(arg debug)"/>
     <arg name="gui" value="$(arg gui)" />
     <arg name="headless" default="$(arg headless)"/>
@@ -38,6 +41,7 @@
     <arg name="task_kind" value="$(arg task_kind)" />
     <arg name="circle_motion_times" value="$(arg circle_motion_times)" />
     <arg name="rqt_gui" value="$(arg rqt_gui)" />
+    <arg name="motion_planning" value="$(arg plannar)" />
   </include>
 
   <!-- 2D Mapping -->
